+++ docs_src/src/guide/changelog.md
@@ -2,6 +2,11 @@
 
 [[toc]]
 
+## v2.10.0
+
+- Added router `Get`, `Post`, `Put`, `Patch`, `Delete` and `Options` methods to register routes directly without having to specify a method string.
+- Added [placeholder](./advanced/localization.html#placeholders) support in regular language lines.
+
 ## v2.9.0
 
 - Added [hidden fields](./basics/database.html#hidden-fields).